+++ src/kinematics/testKinematics.cpp
@@ -1,19 +1,8 @@
 #include "Kinematics.h"
+#include "../util/func.h"
 
 using namespace std;
 using namespace MotionControl;
-
-double rad2deg(double radian)
-{
-	double degree = radian * 180.0f / pi;
-	return degree;
-}
-
-double deg2rad(double degree)
-{
-	double radian = degree * pi / 180.0f;
-	return radian;
-}
 
 int main(int argc, char* argv[])
 {
